Fix memory leak in fight app
The fight app allocates the FightHandles structure twice: once in fight_init()
and again in fight_start(). The first allcation in fight_init() is always
lost, resulting in a memory leak of 104 bytes each time we battle.

The allocation in fight_init() has been removed to correct this.

Currently you can't stash anything in the context->priv pointer in the
init method: the orchard app framework assumes you will return a number
of bytes as a return value from the init method. It will then allocate
space on the app thread stack to use as storage and set context->priv
to point to it.

Using stack space like this may seem clever, but don't do it. Clowns will
eat you.

We want to allocate space from the heap in the start method instead. We
always want to return 0 from the init method to prevent the stack allocation.
But the result is that the orchard_app_thread() function sets context->priv
to NULL. This means the allocation of FightHandles done in the init method
is always lost.

Even if that didn't happen, the context->priv handle would get overwritten
in the start method when FightHandles is allocated a second time.

Also, the redraw_badge() function in the default/badge app has been
updated so that both calls to putImageFile() are protected by the if()
statement. Previously, the second putImageFile() call would always
trigger while the user's health bar was updated while healing, causing
extra traffic to the SD card. This doesn't really hurt, but it does
consume a little more power.