Fix a bug in the firmware updater app
The firmware update app has to block other threads from running before
it loads the updater.bin image into memory. The updater image is loaded
into space that is used for the heap. If the region where the updater
goes is already allocated by a thread, then that thread will likely
crash once its data is overwritten.

To avoid this, we set the updater thread's priority to the max before
trying to load the updater.bin image into memory. This will prevent
any other thread (mainly the gtimer thread) from preempting the updater.