Fix a delay glitch when running the credits app
The credits app does all its work in its start routine. It never gets
to run its event loop. That means if a radio event arrives and triggers
the radio event callback, it could block for up to 5 seconds. This in
turn prevents the main thread from launching the next app until the
5 second timeout expires. To avoid this, we call the radio_event()
funtion to wake it up before the current app thread exits.

Note that we do this after instance->context is set to NULL, so
radio_event() won't do anything except wake up main thread (if it's
asleep).