Fix a bug in the dac thread.
In a last minute change before checking in the dac driver, I broke the
termination conditions of the player loop. We were not actually exiting
on end of file. But since the dac thread has the lowest priority out
of all threads in the system, it was hard to see anything was wrong.
The termination condition is fixed now.