Make the LED code less of a CPU hog
The WS2812 LEDs require pulses to be generated with a specific set of
on/off timings. The code implemented here achieves this using the ledUpdate()
assembly routine that uses a hard coded set of NOP instructions to achieve
the necessary timing. (In effect it uses a manual delay based on the
a specific number of CPU cycles.)

One problem with this approach is that due to the critical nature of the
pulse timing, it's important that the LED thread not be preempted while
generating the pulses. To achieve this, the code guards calls to ledUpdate()
woth chSysLock()/chSysUnlock() which blocks out other threads.

Unfortunately this will also block the DAC thread periodically, which
causes a small amount of choppiness in the audio output.

It turns out that the yse of chSysLock()/chSysUnlock() to guard the entire
ledUpdate() function is overkill. ledUpdate() transmits all three bytes
that represent an LED color, which is 24 bits. However we really only
need to block preemption when sending each individual bit. This drastically
reduces the amount of time that other threads are blocked and avoids the
warble/chop in the audio.

To correct this, the ws2812b_ll.s code has been modified to mask CPU
interrupts only when transmitting individual 0 and 1 bits using the
cspid/cspie instructions. The calls to chSysLock()/chSysUnlock() in
led.c have also been removed, There doesn't seem to be any ill effects
on the LED effects.

Also, the PIT code has been updated to allow the DAC library to turn off
the second PIT interrupt when no audio is playing (to reduce unnecessary
interrupts).