Increase the SPI clock rate for the touch controller
The XPT2046 code was set up to use a very slow SPI clock when talking to
the touch controller. With an amplifier hooked up to the DAC and TPM pins,
there is a noticable sputtering in the speaker which seems to be caused
by either crosstalk of RF interference generated by the touch controller
polling signals.

The clock divisor has been changed to make the SPI bus clock much faster
(now it's in the MHz range, before it was knocked down to the KHz range).
This seems to eliminate much if not all of the sputtering without negatively
impacting the touch controller. (At least, not for my reference boards
with the East Rising displays.)

It's not clear if this will be a problem for the production boards.