Correct the idle timeout calculation
The timeout is expressed in ticks, however the actual ratio of ticks
to seconds may vary depending on the setting of CH_CFG_ST_FREQUENCY. To
get the right value, we have to use the MS2ST() macro to convert the
delay in milliseconds to ticks. Otherwise, if CH_CFG_ST_FREQUENCY is
increased, the idle timeout triggers too soon.