Avoid possible deadlock in userconfig code
The configStart() function initializes config_mutex, then takes it. The
configSave() function also takes it. You don't need to do this, and it
seems in some cases it may cause a deadlock (the main thread blocks).

Taking the mutex in configStart() is not needed, so this has been
removed. (It's still initialized.)

Also, in orchardAppTimer(), we now use MS2ST() instead of US2ST() for
large microsecond waits. For now this works, but with a higher system
tick resolution, the US2ST() macro will calculate an invalid tick count
for any value over about 490000 microseconds.