Add preliminary support for a button event
The SW4 button on the Freedom board is now connected to a sample
interrupt handler. This serves as a template for the tilt sensor (to
allow us to rotate the display when the user turns the badge upside down).

Events are detected on putton presses and releases. Currently a press
plays the Homer WhooHoo! sound effect and a release plays the Doh!
sound effect.