Fix a potential stability issue when resuming threads
It turns out that calling chThdResume() on a thread that isn't actually
sleeping is an error. In extreme cases, it can lead to the scheduler
getting hung. I discovered this by pressing the SW4 button on the Freedom
board over and over very rapidly to exercise the Doh!/Woohoo! audio
playback. Pressin the button quickly enough can sometimes result in
dacPlay() doing a chThdResume() when the dac thread isn't actually
sleeping. If you do this, the scheduler gets jammed: it looks as if the
dac thread ends on the run queue twice, which causes the scheduler to
get stuck in a loop.

To correct this, we now use the ChibiOS thread message facility to wake
up the dac thread (as well as the tpm thread, and the orchard-app radio
event handler). This requires turning on the message feature, which
requires a small amount of additional code, but this fixes the problem
(and also makes the dacPlay() function a little simpler).