Add an app-hackers app
This is is a sample app to demonstrate video playing. It's fairly
primitive, but functional. The intent was to have this be a hidden app
that runs at boot time and plays an randomly selected clip. This may
not be desireable in conjunction with the logo function though as it
may make the boot time too long. It could be set to randomly show a
clip before returning to the "Badge" screen instead, or become the
basis for another app entirely.

Video is displayed in the most primitive way possible: by just loading
full frames from SD card one at a time at a rate of 5 frames per second.
A matching audio track is played in the background via the dac.

In order to make this work, a couple things had to be tweaked:

- The native image display routine now waits one tick between drawing
  image chunks
- The system tick resolution is now 10000

This was required to be able to play audio in the background while
displaying video data. Having the image display function sleep periodically
prevents it from starving out the dac thread and blocking the audio
playback. The system tick resolution had to be increased because the
default of 1000 ticks per second was too slow.

An encode_video.sh script has been added to convert a video file to
static frames and an audio track using ffmpeg.

Right now the app does one of three things:

- Plays a short clip from Hackers (hence the name)
- Plays a short test clip of me
- Shows the FBI logo and plays an audio clib of Agent Gill.

Videos must be stored in a separate directory (all the video frame and
audio files for a given video are collected there). If a video has only
one frame, then just that frame is shown while the audio plays.

For the two video clip cases, the user can touch the screen to end playback
of the video clip.

Video clips should be kept short: audio/video sync is not perfect by
any means, and the longer the video plays, the more it will fall out of
sync with the audio.