Fix graphics when compiling with optimization
The graphics driver behaves erratically when compiling with -O or -O2
instead of -Os. This kind of behavior is usually a sign of slightly
incorrect code.

A few minor changes have been made. After writing to the data register,
you apparently have to do a dummy read to make sure the transfer starts.
In some cases I guess we got lucky when using -Os and things happened to
work, but with compiler optimization turned up, sometimes transfers would
not happen corretly and the screen would be scrambled.