Improve the video and audio encoding utilities
It's possible to make ffmpeg directly output raw video files in
rgb565 formatted-frames, which avoids the need for exploding the
file out into separate .jpg files and processing them individually.
This is much faster and avoids the need for the ImageMagick convert(1)
utility.

Also, ffmpeg can be used to dump the audio track from a video
directly to a WAV file.

The snd16to12.c utility has been updated to handle up to 1024 samples
at a time to make it much faster. (Before it just processed one
sample at a time, which worked but was dirt slow.)

There should be no changed in functionality.