Change optimization back to -Os
The display driver still doesn't work right when using -O2. There's a
problem with the initialization routine: when using -O2, I think a command
is being lost somewhere. I was fooled into thinking it worked because
I had previously run an image built with -Os on the prototype board,
and it set the display configuration correctly, then loaded the image
built with -O2 right after, without doing a hard reset. The screen
retained the proper configuration from the previous boot. But when
running the -O2 image after power-cycling the screen, the display
was garbled.

This will be investigated further, but we need things to be stable
until I can figure out where the problem is.