Add DMA support for SPI controller
I finally figured out the magic for getting the DMA controller to work
in conjunction with the SPI controller. We now use DMA when performing
block read transfers from the SD card. This yields about a 30% improvement
in speed, and ensures smooth video/audio playback.

A DMA module has been added to initialize the controller and handle
DMA completion interrupts. The rcvr_spi_multi() will now queue up the
block read transfer and suspend the current thread until it completes.
This allows other threads to run while the transfer is in progress.

Using the DMA controller for block transfers to the screen may also be
possible, but this has not been implemented yet. It's not clear yet how
much of a performance gain this would provide.

Note that when building the firmware updater, rcvr_spi_multi() uses the
slower synchronous read code to keep the updater size to a minimum.