Make a few minor tweaks
Set the asynchronous bit when initializing DMA transfers. I don't think
this makes any difference in performance, but it doesn't seem to hurt
either, and the mnaual implies it might be correct to enable it.

Also, shave a few instructions out of xchg_spi() in mmc_spi_lld.c. This
may mak a microscopic amount of difference in SD card I/O performance,
and it also doesn't seem to hurt.