Avoid a possible race in app radio event processing
Instead of pausing the main thread when handing off a packet to the app
thread, just copy the packet and schedule the app thread immediately. It's
possible the main thread might become unblocked unexpectedly if the radio
interrupt fires. We may drop a packet if one arrives before the current
one is processed, but such is life.

Also, disable the tick event debug messages for now and clean up some of
the other debug message formatting in the fight app.