Make a few more cleanups
It turns out we don't need to increase the thread priority in app-hackers
when playing video. I did this early on when debugging video playback and
never removed it, but with the current implementation it doesn't seem to
be necessary.

Also, when performing DMA, it turns out it's not necessary to clear and
re-set the enable bit in the SPI control register after all. I'm not sure
why I felt it was needed before. It's possible that I hadn't yet stumbled
onto the correct sequence for initiating a DMA transfer, but experimentation
with the current code shows that this can be removed from both the MMC
and display drivers and SPI DMA transfers still work fine. Removing this
saves us a few instructions per transfer, so we may as well.