Fix the firmware updater again, add new video feature
Now that we use DMA for SD card reads, the hack of raising the updater
thread's priority to the max won't work anymore: when performing DMA,
we sleep the current thread, and when we're asleep, the gtimer thread
might run.

So to really make this work right, we have to directly go out and really stop
any events that can hurt us. The updater code now terminates the gtimer
thread and the LED thread, and turns off radio events so that we don't
get interrupted by the main thread. (The DAC and TPM threads should be
idle so we don't have to worry about them.) This should make the firmware
updater work correctly in all cases.

Also, the video module has been updated to support a second video playback
mode. In addition to videoPlay(), there is now a videoWinPlay() which will
play the video without upscaling (i.e. in 128x96 pixel format), positioned
at any given X and Y coordinates on the screen that the caller specifies.
This can allow small video effects to be added to other apps without
overwriting the entire display. Calling videoWinPlay() with x and y set
to 0 causes video playing to use the whole screen.