Reclaim some RAM
The awful, no-good Caesar cipher command uses a static output buffer.
This buffer is placed in the BSS, meaning it constantly uses 255 bytes
of RAM, even if we never use the command.

The code has been modified to allocated the buffer from the heap instead.
That way the memory is only allocated when it's needed instead of hogged
forever. This frees up the 255 bytes.