NETWORK BREAKINNG CHANGE
Add a field to implement a very naive datetime "virus" like protocol
for sending the clock along with the ping packets. This is a change ato
the network protocol itself and will break fight and ping if you do
not update all boards.

Update the LED subsystem. When the SD card check fails, cycle all of
the lights from red to green to blue so that we can use the
screen-not-installed state as an LED debugging state.