Set PTA4 and PTA19 to enable internal pullup
With the rev2 prototype, the SWENTER/PTA4 button generates spurious
events when you touch the pads. This is because it needs a pullup.
Fortunatelly, we don't have to use an external one: we can just enable
the internal one in the chip. For some reason I had set PTC1, PTC2 and
PTC4 this way, but not PTA4 or PTA19. This has been fixed.