Implement the fully-working 'clock virus'. If any badge in the network executes:
config set rtc XXXXX

where XXXXX is a current and valid unix time_t, other badges will
redistribute the RTC data. Once a non zero RTC is received, the clock
will get set in that badge. Data is not maintained across resets, so
resetting a badge next to someone with an accurate clock will get you
the "right" time.

The major failure mode with this is that if you set the time wrong
once, everyone will redistribute the wrong time.

A 'timelord' value that forcibly resets the clock might be fun here,
or checking for a set of boundaries (i.e. only works on defcon dates.)

I don't know, this is all experimental.