Add a software workaround for broken rev2 proto badge B
Of the two rev2 prototype badges, badge B has a faulty solder connection
on pin 49 of the MCU. This means we're not getting any radio interrupts.
A software workaround has been added to hijack one of the TPM timers to
poll the radio RX handler periodically.

The code for this is only compiled with built with -DREV2_RADIO_WAR, which
is off by default. To enable the workaround, the build option for it can
be uncommended in the Makefile.

This is only needed for this faulty badge and is just a hack so that we
can still do some testing with it.