Add event handling for joypad and tilt sensor
Interrupt/event handling for the up/down/left/right joypad buttons has
been added, as well as support for the tilt sensor.

To enable them, uncomment the ENABLE_JOYPAD and ENABLE_TILT_SENSOR options
in the Makefile.

This are off by default because: a) the Freescale/NXP reference board
doesn't have them and it has the pins wired differently and will freak
out when the interrupts are enabled, and b) the tilt sensor seems to
behave very badly on the prototype board. It may need some software
de-bouncing before it is useful. Right now it's just annoying.