Connect joypad to apps
The keyEvent type has been reworked so that it now handles joypad events.
When a key is pressed, apps will get an event notification notifying
them which key triggered the event and if it was a press/release event.

The Homer Doh/Woohoo sound effect has also been removed. There's still
a message printed to the console, but this is just for debug purposes
and will be removed eventually too.