Fix up joypad button
Apparently PTA19 (up button) doesn't need the internal pullup enabled.
There seems to already be 3volts on it. I'm sure why. But having both
that _and_ the internal pullup enabled causes spurious double-triggering
of "up" button press and release events.

We need to confirm this is the same on both rev2 proto boards and not
just a fluke with one.