Add a newer Mortal Kombat song
The previous Mortal Kombat recording was overamplified and had a "blown out"
sound to it. Also, sox had trouble converting the original MP3 (it claimed
to lose sync) which caused some skips in the playback. The new version is
ripped from the Mortak Kombat theme music video on Youtube and sounds a
little better.