Fix a bug in the DAC module
If we're playing videos, we don't want to allow the DAC thread to also
play music at the same time, as this confuses the thread logic. If
we detect that the video app is running or that someone has already
queued up samples to play via dacSamplesPlay(), we abort any attempt
to play a tune sent to us via dacPlay().