Implement a less obtrusive way of blocking audio during video playback
The previous attempt to keep audio from playing via the dac thread during
video playback was a little awkward in that it required making the DAC
low-level driver code aware of the Orchard app code. We can use a simpler
method which is to check to see if dacBuf is NULL instead. If it isn't,
that means the video code is using it.