NETWORK BREAKING CHANGE (network config version now "13")
There is now a p_type of notype, which is the state you are in before
you choose a character. If you are in this state you will always have
in_combat == 1 so no one can start a fight with you.

I've also modified the boot sequence. If you have no config it goes:

boot screens -> enter name -> choose type.

If you attempt to reset the board halfway through, we will force you
back to the name or choose type screens.

Aside from that, you get to choose your type once and only once. Then
we hold you to it!