Fix a bug in the DAC module
With the new clicks generated by joypad button presses, the credits
app wouldn't play the mario tune. Due to something I did to support
dac looping, there was a bit of a race with the dac thread. It would
set the fname pointer to NULL when exiting. If you schedule a tune to
play, then schedule another before the first one quits, the fname will
be set to NULL after the new tune filename has been passed to dacPlay(),
but before we actually get to play it.