Make a few small updates and improvements
The click effect is now played from orchard-app,c for both joypad and
ugfx events. This saves apps from having to generate the effect manually.
Also, things like the list widget and ugfx buttons now generate click
noises on their own. The keyboard UI generates clicks for typed letters
too.

Also, there was still a subtle logic bug in the DAC player. If you went
to the keyboard UI and kept typing until you hit the end of the buffer,
it would try to play a click and then a homer "Doh!" sound. However it
actually ended up playing two "Doh!" sounds instead. The reason is
that if you call dacPlay() twice in a row from the same thread, the
first call would start the dac thread, but it would get preempted
immediately by the app thread again. Then the app thread would call
dacPlay() again before the first sound had a chance to play and override
the fname pointer with a new filename. So the dacPlay() to play the
click would actually play a Doh!. This has been corrected.

Also, the decimal to roman numeral converter was using RAM for its
convertion table instead of putting them in ROM.

The video app doesn't need to reset the PIT frequency if the auto-adjust
hack is not enabled, so we can conditionalize that.

Lastly, the fire LED effect has an off-by-one error that leads to
memory corruption. This has been fixed, but it the effect seems to
not work that well either.