Add touch screen calibration app
An app-calibration.c app has been added which uses the uGFX calibration
GUI to allow the user to customize the touch screen calibration values.
The values are saved to flash in the userconfig structure.

The factory default values will be used if no user-specified values
are present in the flash.

Also, the width of the scrollbar in the list widget has been widened
to make it easier for users to hit it. Before it was a bit too skinny.