Add another firmware update fix
Some linker magic has been added to ensure that the data structures
and buffers used by the FatFs code are positioned as close to the bottom
of memory as we can manage. This ensures they won't end up being clobbered
when reading the firmware updater from SD card.

The image has grown large enough that the fs_mutex mutex was now being
positioned past the start of UPDATER_BASE, which made the firmware update
process fail.