Add a LED sign app
This is based on the scroll_lld code. The user can enter a short message
which will be scrolled across the screen using an LED font. Tapping the
screen exits.

The font is pre-rendered as RGB files which are loaded one by one to
match each character in the string. This allows the app itself to be
as small as possible to preserve flash.