Fix issue #57
The sts variable in the ledsign_event() function was uninitialized.
If the user entered an empty string for the sign, the display code
would get stuck in an infinite loop.

Initialize sts to a non-zero value so to signal error by default.
This will be changed only if the scroll function returns success when
rendering a character. (It returns non-zero when the user taps the
screen to exit.) This way if there's an empty string where strlen(str)
is zero, the for() loop never calls the scroll function and we exit
right away.