Adjust fectory reset feature a little
On the Freescale/NXP reference boards, the UP and DOWN signals are
always pulled low, which fools the userconfig code into thinking that
the user always wants a factory wipe at startup. To fix this, use
the UP and DOWN buttons only when joypad support is enabled, and use
the enter/select button when it isn't.