Fix a bug in the video app
If the SD card is not present, the video app will crash the system you
launch it. If the SD card is missing, the selection list is never
populated and context->priv id NULL. But the video_exit() function
doesn't check for this, and it tries to pass a bogus pointer to
chHeapFree() as a result. We test for the NULL pointer and exit
gracefully now.