Add a small spectrum display app
This app uses the RSSI readings from the radio to generate a simple
spectrum display. I'm not sure it should be put in the final production
built or not since it usefulness is questionable. We don't really have
enough speed or memory to sample the spectrum effectively. Single radio
packet transmits from the badge don't show up reliably because the packet
may have already disappeared by the time we sweep to the right place.

The app works by reading the RSSI values at 320 intervals of a selected
span size, centered on the main carrier frequency. Currenly the span size
is 500KHz, which yields a coverage of 841MHz to 1001MHz.