Make a couple of updates to the SD/MMC driver
There is yet another possible way to hose the firmware update process.
The timer variables in the MMC module must be kept further down in memory.
I found that it this isn't done, the Timer1 and Timer2 vaiables might
end in the space where we try to load the updater.bin file, which corrupts
the code.

Also, we now use multiblock reads by default for all cases. This can
be turned off by turning off the MMC_FORCE_MULTIBLOCK_READ macro in
mmc_spi_lld.c. It seems to work ok and seems to yield better results
with the spotty 16GB SD card that I have.