Fix SD card detection
Previously, when fixing the firmware updater, the Stat variable in
mmc_spi_lld,c was tagged to go in the .fsdata section. The problem is
that the .fsdata section was actually in the bss section, which meant
its initialized value was being zeroed. This meant we always thought
the SD card was present.

This has been fixed as followed:

- The .fsdata section has been moved to the .data section
- A .fsbss section has been created too
- All previously tagged variables from the .fsdata are now on .fsbss
- Stat has been moved to .fsdata