Fix a backspace bug in the keyboard UI
Scrolling back up a line doesn't work. The console widget doesn't draw in the
last column because it thinks there's not enough room for a character there.
(There actually is, but whatever.) This means we have to skip back an
extra space when backspace brings us back up a line.