Add a way to order the app list
The orchard_app() macro now includes a "prio" argument which can be
used to set the order in which an application appears in the orchard_app_list
array. Currently the order is determined somewhat arbitrarily by the linker.
We would like certain apps to appear first in the list so that they remain
in the same place when new apps are added.

The prio field is a number which is incorporated into the section name
used when tagging the app's orchard_app structure as part of the linker
set. The default app (Badge) is currently set to 0, which is the
highest priority. Some apps such as setup, fight and credits use values
1, 2 and 3. The remaining apps use priority 9999, which means we don't
care yet what order they're in. This can be refined later.

Hidden apps can always have priority 9999 since they never show up in
the launcher menu anyway.

This does not add any additional size to the image since it just controls
the section names.