Clean up tetris a little
The tetris playfield is actually too big to fit on the streen in landscape
mode. To overcome this, we temporarily switch to landscape mode while the
game is running, and switch back when the game is over.

The "Paused!" message was not being erased properly. (One row of pixels
was left behind on the screen.) This has been fixed.

A one second pause has been added before exiting.

Also, spurious ^Ms in the tetris.h file have been removed.