Add photo frame app
This is a basic photo frame app. Photos should be placed in the
photos subdirectory on the SD card. The app will side-scroll them
onto the screen at regular intervals until the screen is tapped
to exit.

There should be some default photos but there aren't any yet. This
is TBD.