Clean up the music app a little
There was a small bug in the spectrum drawing code: sometimes the top-most
line of a column would end up drawn in green instead of red.

Also, use of flowing point numbers has been removed. This reduces the
size of the object code a little to save space in flash.