Restore our frame rate to 15FPS
Remove all of the sleeps from the VS screen so that we can yield when
packets come in

Rewrite the protocol handler to work on a countdown timer instead of
overriding the app thread's timer. This will allow us to service the
network retries at 500mS +/- some value and do smooth animations.