If you kill Caesar, show "ET TU, USERNAME?" - thanks to Katbret for the idea.
Add a deepcopy feature so that we update user stats on battle ack. We
had a strange issue where we would not track HP correctly, dropping
hit points and this would completely screw up the fight progress. May
need to add this copy to NEXTROUND as well.

Add the caesar and senator unlocks, I forgot them.

Add vary-play, to play a random sound on hits