Fix an audio playback bug in the credits app
If the credits app plays all the way through to the end, it should stop
the music playing when it exits, but it was not. This looks like something
that was broken when the scrolling code was factored out into a separate
module.

If the user exits manually, the sound playback stops because the click
effect overrides the music, but if the app exits on its own, nothing
stops the music from playing until either the song runs out or something
triggers another sound. Now the credits app does an explicit dacPlay(NULL)
when it reaches the end of the credit scroll.