Fix a memory leak in the fight app
The arrow button widgets were being created in each round, but were
not being destroyed. This meant we leaked 48 bytes of memory for each
arrow button in each round.

Now the arrow button widgets are released as soon as a move is selected.

One problem with this is that calling gwinDestroy() on thw widdget causes
uGFX to immediately make it invisible, which in this case means drawing
over the arrows and deleting them, along with part of the opponent avatar.
To defeat this, gwinDestroy() has been modified to not call
gwinSetVisible(gh, FALSE). If you really want this behavior, you can
always call this before calling gwinDestroy() manually, but we don't
want it. This doesn't seem to break any other UI elements.