Refine the firmware update process.
The app-verify-update functionality has been merged into app-update itself.
There's no reason to have a separate app for the confirmation screen.

Also, the firmware updater is performs some sanity checking of the image.

The linker script now reserves a 16-byte region right after the flash
security field for storing a magic number and a version value. The
version is the build time, in UNIX "seconds since the epoch" form.
We reserve 64-bits for the the build time so we don't roll over in 2038
(har har).

The firmware_version structure in app-update.c uses an attribute tag
to ensure it's stored in the reserved region. This allows us to know
exactly where to find it in the binary firmware image: it's always at
offset 0x420.

If app-update finds a valid updater image and badge.bin image, it
will read the magic number and version field. If the magic number
isn't valid, the update aborts. If the timestamp in the badge,bin
image isn't newer than the currently running firmware, the update
aborts.

This prevents you from re-flashing the same image over and over,
and will stop anybody from flashing the badge,bin file that was
accidentally shipped on the SD card.