Fix formatting issue in error message
The "invalif firmware" message's second line was too long and would
overflow into the next line. Also, the magic numbers are 64-bit values,
but there is no gwinPrintf() format that supports printing 64-bit hex
numbers. Right now we only use the lower 32 bits of the magic number
anyway, so we use a cast to get the values to display correctly as
32-bit hex numbers.