Claw back some more space after the last rount of changes
The temperature calibration command code was using strtol() instead of
strtoul(). The latter function is provided in cut down form to avoid
using the one in libc, which is quite large. Using strtol() dragged
it back in again, which consumed over 700 bytes for no good reason.
The calibration value is an unsigned integer so using strtol() wasn't
really needed anyway.

Also, code was added to app-default to save the gdispGetHeight() and
gdispGetWidth() values so that these routines only needed to be called
once, but the saved gdispGetWidth() value was then being ignored. Fixing
this saved a few more bytes.

Lastly, the calibration offset is now applied to in radioGetTemperature()
so that it only needs to be set in one place instead of doing two
calibration steps.