Trimp down the fight logic to save more space
With a few logic optimizations we can save another 200 bytes of flash space.

When checking keypresses, checking for the key _and_ checking for the
pressed state each time is redundant. The fight logic only cares about
keypress events, not release events. So we can just check for the keyPress
flag once and then decode which button was pressed.

In the grant code, sendGamePacket() was seemingly called twice to send
a grant: once for each grant type in each statement case, and then again
at the end. This is also redundant.

When choosing what audio sample to play, instead of calling dacPlay()
once in each case of a large switch statement, we can save some code by
just choosing the audio effect name and then calling dacPlay() once
and the end.

Lastly, some of the drawing was optimized. When instantiating the
attack arrow buttons, they are invisible, yet we were assigning
the text field for each widget (even though it would never appear).