Make another minor optimization to save some space
It turns out that to divide by 2, it's cheaper (in terms of emitted code)
to do:

foo >> 1

instead of:

foo / 2

even though they have the same effect. It's possible the compiler may be
trying to call a math intrinsics routine to do the division, whereas
doing the shift saves a function call. In any case, this saves about
50 or 60 bytes.