Spruce up the dialer a little
Support for the newer UI has been merged in. Support for the blue box tones
is present but is #ifdef'ed off for now. The frequencies of 1500Hz and 1700Hz
yield some tones which do not play correctly without increasing the sample
rate, but doing that also increases the amount of RAM required to hold
one wave's worth of samples to the point where we don't have enough free
memory. If I can think of a way to turn avoid this I may turn it back on.

Still to be done: figure out a way to avoid the keyclick sound effect
when pressing a button.