Fix some UI issues
The change to handle both button up and button down events broke a few things.

The video, scroll and photos apps need to handle checking for screen taps
a little differently. We want to make sure we trigger only on mouse down
events, not mouse up events (otherwise we might exit immediately as soon
as the user takes their finger off the screen after tapping to select
an app).

Also, we now play a touch tone sequence on boot. For now the sequence
is 867-5309. This will change soon.