Fix a minor nit in the video player
Previously the test to exit the player loop was changed to detect
mouse down events rather than just any touch event. The same test was
also done a second tine to decide what value to return to the caller,
but that test was still checking for touch events. It's now been
updated to also check for mouse down events.