Fix a search and replace botch
When the "textextra" feature was introduced, somehow some instances of the
string "extra" got expanded into "textectra", including the "extra" argument
to the write_pixel() and draw_lines() functions. This has been fixed.

This doesn't change any behavior. IT just fixes the argument/variable
name. The resulting code is unchanged.