Fix the firmware updater once again
If you use the GNU ARM Eclipse toolchaim from September 2016 which has
this version string:

gcc version 5.4.1 20160919 (release) [ARM/embedded-5-branch revision 240496] (GNU Tools for ARM Embedded Processors)

Then the "ch" structure can end up in the area of RAM where the firmware
updater is loaded. This means when we try to load updater.bin into RAM,
we overwrite the head of the ChibiOS thread list.

To fix this, we need to tag it so that the linker puts it in the "fsbss"
section to keep it out of the way.